requirements.txt
# Python dependencies for CUDA_CONV_CNN
numpy>=1.24
pybind11>=2.11
tqdm
matplotlib
pillow

# Data loading (for CIFAR-10)
torch
torchvision

# Visualization
tensorboard
